backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
motor==3.3.2
beanie==1.23.1
pymongo==4.6.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# API & Validation
email-validator==2.1.0
httpx==0.25.2
python-dateutil==2.8.2

# File handling
python-magic==0.4.27
pillow==10.1.0
aiofiles==23.2.1

# PDF Generation
reportlab==4.0.7
pypdf==3.17.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Google Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0

# WhatsApp Integration (via HTTP)
httpx==0.25.2

# Timezone & Scheduling
pytz==2023.3
apscheduler==3.10.4
croniter==2.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# CORS & Security
python-multipart==0.0.6
itsdangerous==2.1.2